+++ packages/gatsby-source-filesystem/src/gatsby-node.js
@@ -129,12 +129,15 @@ See docs here - https://www.gatsbyjs.org/packages/gatsby-source-filesystem/
   const watcher = chokidar.watch(pluginOptions.path, {
     ignored: [
       `**/*.un~`,
+      `**/.DS_Store`,
       `**/.gitignore`,
       `**/.npmignore`,
       `**/.babelrc`,
       `**/yarn.lock`,
+      `**/bower_components`,
       `**/node_modules`,
       `../**/dist/**`,
+      ...(pluginOptions.ignore || []),
     ],
   })
 
@@ -150,6 +153,8 @@ See docs here - https://www.gatsbyjs.org/packages/gatsby-source-filesystem/
         currentState = fsMachine.transition(currentState.value, `EMIT_FS_EVENT`)
         createNode(fileNode)
       }
+
+      return null
     })
     return fileNodePromise
   }