+++ packages/gatsby/src/internal-plugins/query-runner/file-parser.js
@@ -95,38 +95,92 @@ async function findGraphQLTags(file, text): Promise<Array<DefinitionNode>> {
          */
         const documentLocations = new WeakMap()
 
+        const extractStaticQuery = taggedTemplateExpressPath => {
+          const { ast: gqlAst, text, hash, isGlobal } = getGraphQLTag(
+            taggedTemplateExpressPath
+          )
+          if (!gqlAst) return
+
+          if (isGlobal) warnForGlobalTag(file)
+
+          gqlAst.definitions.forEach(def => {
+            documentLocations.set(
+              def,
+              `${taggedTemplateExpressPath.node.start}-${def.loc.start}`
+            )
+            generateQueryName({
+              def,
+              hash,
+              file,
+            })
+          })
+
+          const definitions = [...gqlAst.definitions].map(d => {
+            d.isStaticQuery = true
+            d.text = text
+            d.hash = hash
+            return d
+          })
+
+          queries.push(...definitions)
+        }
+
         // Look for queries in <StaticQuery /> elements.
         traverse(ast, {
-          TaggedTemplateExpression(path) {
-            if (
-              (`descendant of query`,
-              path?.parentPath?.parentPath?.node?.name?.name !== `query`)
-            ) {
+          JSXElement(path) {
+            if (path.node.openingElement.name.name !== `StaticQuery`) {
               return
             }
-            if (
-              path.parentPath?.parentPath?.parentPath?.node?.name?.name !==
-              `StaticQuery`
-            ) {
-              return
-            }
-            const { ast: gqlAst, text, hash, isGlobal } = getGraphQLTag(path)
-            if (!gqlAst) return
 
-            if (isGlobal) warnForGlobalTag(file)
+            // astexplorer.com link I (@kyleamathews) used when prototyping this algorithm
+            // https://astexplorer.net/#/gist/ab5d71c0f08f287fbb840bf1dd8b85ff/2f188345d8e5a4152fe7c96f0d52dbcc6e9da466
+            path.traverse({
+              JSXAttribute(jsxPath) {
+                if (jsxPath.node.name.name !== `query`) {
+                  return
+                }
+                jsxPath.traverse({
+                  // Assume the query is inline in the component and extract that.
+                  TaggedTemplateExpression(templatePath) {
+                    extractStaticQuery(templatePath)
+                  },
+                  // Also see if it's a variable that's passed in as a prop
+                  // and if it is, go find it.
+                  Identifier(identifierPath) {
+                    if (identifierPath.node.name !== `graphql`) {
+                      const varName = identifierPath.node.name
+                      let found = false
+                      traverse(ast, {
+                        VariableDeclarator(varPath) {
+                          if (
+                            varPath.node.id.name === varName &&
+                            varPath.node.init.type ===
+                              `TaggedTemplateExpression`
+                          ) {
+                            varPath.traverse({
+                              TaggedTemplateExpression(templatePath) {
+                                found = true
+                                extractStaticQuery(templatePath)
+                              },
+                            })
+                          }
+                        },
+                      })
+                      if (!found) {
+                        report.warn(
+                          `\nWe were unable to find the declaration of variable "${varName}", which you passed as the "query" prop into the <StaticQuery> declaration in "${file}".
 
-            gqlAst.definitions.forEach(def => {
-              documentLocations.set(def, `${path.node.start}-${def.loc.start}`)
-              generateQueryName({ def, hash, file })
+Perhaps the variable name has a typo?
+
+Also note that we are currently unable to use queries defined in files other than the file where the <StaticQuery> is defined. If you're attempting to import the query, please move it into "${file}". If being able to import queries from another file is an important capability for you, we invite your help fixing it.\n`
+                        )
+                      }
+                    }
+                  },
+                })
+              },
             })
-
-            const definitions = [...gqlAst.definitions].map(d => {
-              d.isStaticQuery = true
-              d.text = text
-              d.hash = hash
-              return d
-            })
-            queries.push(...definitions)
+            return
           },
         })
 
@@ -145,7 +199,11 @@ async function findGraphQLTags(file, text): Promise<Array<DefinitionNode>> {
                     def,
                     `${innerPath.node.start}-${def.loc.start}`
                   )
-                  generateQueryName({ def, hash, file })
+                  generateQueryName({
+                    def,
+                    hash,
+                    file,
+                  })
                 })
 
                 queries.push(...gqlAst.definitions)