+++ packages/gatsby-source-contentful/CHANGELOG.md
@@ -3,6 +3,72 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.2"></a>
+
+## [2.0.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.1...gatsby-source-contentful@2.0.2) (2018-10-03)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+<a name="2.0.1"></a>
+
+## [2.0.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.1-rc.9...gatsby-source-contentful@2.0.1) (2018-09-17)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+<a name="2.0.1-rc.9"></a>
+
+## [2.0.1-rc.9](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.1-rc.8...gatsby-source-contentful@2.0.1-rc.9) (2018-09-11)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+<a name="2.0.1-rc.8"></a>
+
+## [2.0.1-rc.8](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.1-rc.7...gatsby-source-contentful@2.0.1-rc.8) (2018-09-11)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+<a name="2.0.1-rc.7"></a>
+
+## [2.0.1-rc.7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.1-rc.6...gatsby-source-contentful@2.0.1-rc.7) (2018-09-11)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+<a name="2.0.1-rc.6"></a>
+
+## [2.0.1-rc.6](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.1-rc.5...gatsby-source-contentful@2.0.1-rc.6) (2018-09-11)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+<a name="2.0.1-rc.5"></a>
+
+## [2.0.1-rc.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.1-rc.4...gatsby-source-contentful@2.0.1-rc.5) (2018-09-07)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+<a name="2.0.1-rc.4"></a>
+
+## [2.0.1-rc.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.1-rc.3...gatsby-source-contentful@2.0.1-rc.4) (2018-09-05)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+<a name="2.0.1-rc.3"></a>
+
+## [2.0.1-rc.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.1-rc.2...gatsby-source-contentful@2.0.1-rc.3) (2018-08-31)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+<a name="2.0.1-rc.2"></a>
+
+## [2.0.1-rc.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.1-rc.1...gatsby-source-contentful@2.0.1-rc.2) (2018-08-29)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
+<a name="2.0.1-rc.1"></a>
+
+## [2.0.1-rc.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.1-rc.0...gatsby-source-contentful@2.0.1-rc.1) (2018-08-29)
+
+**Note:** Version bump only for package gatsby-source-contentful
+
 <a name="2.0.1-rc.0"></a>
 
 ## [2.0.1-rc.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-contentful/compare/gatsby-source-contentful@2.0.1-beta.18...gatsby-source-contentful@2.0.1-rc.0) (2018-08-21)